post request for vending machine done
