+++ backend/index.ts
@@ -2,7 +2,7 @@ import 'dotenv/config';
 import { Hono } from 'hono';
 import { cors } from 'hono/cors';
 import { mapRoutes } from './api/mapRoutes.js';
-import { database } from './api/database.js';
+import { vendMachine } from './api/vendingMachineDB.js';
 import { serve } from '@hono/node-server';
 
 const app = new Hono();
@@ -12,7 +12,7 @@ app.use('*', cors());
 
 // Register map routes under `/api`
 app.route('/api', mapRoutes);
-app.route('/api', database);
+app.route('/api', vendMachine);
 
 // Start server
 const PORT = Number(process.env.PORT) || 3001;
